pyproject.toml
[tool.poetry]
name = "service_approval"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "0.62.0"
fastapi-utils = "0.2.1"
uvicorn = "0.12.3"
requests = "2.24.0"
fastapi_sqlalchemy = "0.2.1"
psycopg2 = "2.8.6"
python-json-logger = "0.1.11"
pilot-platform-common = "0.0.25"

[tool.poetry.dev-dependencies]
pytest = "6.2.5"
pytest-env= "0.6.2"
pytest-asyncio = "0.17.2"
pytest-dependency = "0.5.1"
requests-mock = "1.9.3"
mock = "4.0.3"
sqlalchemy-utils = "0.38.2"
testcontainers = "3.4.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
